Run linting with golangci, small code reorganisation
