+++ .gitignore
@@ -2,5 +2,5 @@
 .idea/
 .vpass
 _build/out/*
-out/*
+dist/*
 vendor/
