+++ .travis.yml
@@ -1,15 +1,12 @@
 language: go
 
 go:
-  - "1.12"
-  - "1.13"
   - "1.14"
   - stable
 
 install:
   - mkdir -p $GOPATH/bin
   - go get github.com/go-playground/overalls
-  - go get golang.org/x/lint/golint
   - GO111MODULE=on go mod vendor
 
 script:
