--- doc.go
@@ -1,6 +1,0 @@
-/*
-Package kandalf is RabbitMQ to Kafka bridge.
-
-For a full guide visit https://github.com/hellofresh/kandalf
-*/
-package kandalf
